# api/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "🐕API DogWalker",
    "description": "➡️Documentação gerada automaticamente pelo módulo <a href='https://github.com/davibaltar/swagger-autogen' target='_blank'>swagger-autogen</a>."
  },
  "host": "backdogwalker.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "access-token",
      "description": "Token de Acesso gerado após o login"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/passeadores/": {
      "get": {
        "tags": [
          "Passeadores"
        ],
        "description": "Endpoint que retorna os passeadores em um raio de 20Km da latitude e longitude informados",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "type": "string"
          },
          {
            "name": "lng",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      },
      "post": {
        "tags": [
          "Passeadores"
        ],
        "description": "Endpoint que inclui um novo passeador",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "type": "string"
          },
          {
            "name": "lng",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Criado"
          },
          "202": {
            "description": "Aceito"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "403": {
            "description": "Acesso Negado"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/api/passeadores/{id}": {
      "get": {
        "tags": [
          "Passeadores"
        ],
        "description": "Endpoint que retorna os dados do passeador filtrando pelo id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      },
      "put": {
        "tags": [
          "Passeadores"
        ],
        "description": "Endpoint que permite alterar os dados do passeador pelo id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "lat",
            "in": "query",
            "type": "string"
          },
          {
            "name": "lng",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Criado"
          },
          "202": {
            "description": "Aceito"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "403": {
            "description": "Acesso Negado"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      },
      "delete": {
        "tags": [
          "Passeadores"
        ],
        "description": "Endpoint que permite excluir um passeador filtrando pelo id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "202": {
            "description": "Aceito"
          },
          "400": {
            "description": "Solicitação Inválida"
          }
        }
      }
    },
    "/api/passeadores/nome/{nome}": {
      "get": {
        "tags": [
          "Passeadores"
        ],
        "description": "Endpoint que retorna os dados do passeador filtrando por parte do nome",
        "parameters": [
          {
            "name": "nome",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/api/prestadores/": {
      "get": {
        "tags": [
          "Prestadores"
        ],
        "description": "Endpoint para obter todos os Prestadores de Serviço do sistema.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Listagem dos prestadores de serviço obtida com sucesso",
            "schema": {
              "$ref": "#/definitions/Prestadores"
            }
          },
          "500": {
            "description": "Erro ao obter a listagem dos prestadores",
            "schema": {
              "$ref": "#/definitions/Erro"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Prestadores"
        ],
        "description": "Endpoint que inclui um novo prestador",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Listagem dos prestadores de serviço obtida com sucesso",
            "schema": {
              "$ref": "#/definitions/Prestadores"
            }
          },
          "201": {
            "description": "Criado"
          },
          "202": {
            "description": "Aceito"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "403": {
            "description": "Acesso Negado"
          },
          "500": {
            "description": "Erro ao obter a listagem dos prestadores",
            "schema": {
              "$ref": "#/definitions/Erro"
            }
          }
        }
      }
    },
    "/api/prestadores/id/{id}": {
      "get": {
        "tags": [
          "Prestadores"
        ],
        "description": "Endpoint que retorna os dados do prestador filtrando pelo id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/api/prestadores/razao/{razao}": {
      "get": {
        "tags": [
          "Prestadores"
        ],
        "description": "Endpoint que retorna os dados do prestador filtrando por parte da Razão Social",
        "parameters": [
          {
            "name": "razao",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/api/prestadores/{id}": {
      "put": {
        "tags": [
          "Prestadores"
        ],
        "description": "Endpoint que permite alterar os dados do prestador pelo id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Listagem dos prestadores de serviço obtida com sucesso",
            "schema": {
              "$ref": "#/definitions/Prestadores"
            }
          },
          "201": {
            "description": "Criado"
          },
          "202": {
            "description": "Aceito"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "403": {
            "description": "Acesso Negado"
          },
          "500": {
            "description": "Erro ao obter a listagem dos prestadores",
            "schema": {
              "$ref": "#/definitions/Erro"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Prestadores"
        ],
        "description": "Endpoint que permite excluir um prestador filtrando pelo id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "202": {
            "description": "Aceito"
          },
          "400": {
            "description": "Solicitação Inválida"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Erro": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "example": "Erro gerado pela aplicação"
        },
        "msg": {
          "type": "string",
          "example": "Mensagem detalhando o erro"
        },
        "param": {
          "type": "string",
          "example": "URL que gerou o erro"
        }
      }
    },
...
    "Prestadores": {
      "type": "object",
      "properties": {
        "cnpj": {
          "type": "string",
          "example": "33719377000156"
        },
        "razao_social": {
          "type": "string",
          "example": "AIMEEVET SERVICOS VETERINARIOS EIRELI"
        },
        "nome_fantasia": {
          "type": "string",
          "example": ""
        },
        "ddd_telefone_1": {
          "type": "string",
          "example": "1176018050"
        },
        "ddd_telefone_2": {
          "type": "string",
          "example": ""
        },
        "cep": {
          "type": "string",
          "example": "04776150"
        },
        "logradouro": {
          "type": "string",
          "example": "MARIO DE OLIVEIRA"
        },
        "numero": {
          "type": "string",
          "example": "421"
        },
        "municipio": {
          "type": "string",
          "example": "ITU"
        },
        "bairro": {
          "type": "string",
          "example": "JARDIM PAULISTA II"
        },
        "uf": {
          "type": "string",
          "example": "SP"
        },
        "porte": {
          "type": "string",
          "example": "MICRO EMPRESA"
        },
        "natureza_juridica": {
          "type": "string",
          "example": "Empresa Individual de Responsabilidade Limitada (de Natureza Empresária)"
        },
        "cnae_fiscal": {
          "type": "number",
          "example": 7500100
        },
        "cnae_fiscal_descricao": {
          "type": "string",
          "example": "Atividades veterinárias"
        },
        "data_inicio_atividade": {
          "type": "string",
          "example": "2019-05-24"
        },
        "descricao_situacao_cadastral": {
          "type": "string",
          "example": "ATIVA"
        },
        "localizacao": {
          "type": "object",
          "properties": {
            "type": {
              "type": "string",
              "example": "Point"
            },
            "coordinates": {
              "type": "array",
              "example": [
                -23.28861,
                -47.31145
              ],
              "items": {
                "type": "number"
              }
            }
          }
        }
      }
    },
...
  }
}